backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.17
supabase==2.10.0
google-generativeai==0.8.3
pillow==11.0.0
pillow-heif==0.20.0
httpx==0.27.0
python-dotenv==1.0.1
pytest==8.3.4
eval-type-backport==0.2.0
PyJWT==2.10.1
